Feature/khronos active window (#321)
* start on active window interface

* (wip) reorg and rename reconstruction module pieces

* fix compilation and add graph update field

* rename InputQueue to MessageQueue

* update reconstruction shutdown logic

* add num blocks archived to return

* add generic update class for putting nodes in graph

* add graph config declaration

* rename FrontendModule

* rename frontend queue method

* udpate config field name

* reshuffle map and map config

* push layer flags to config in VolumetricMap

* update volumetric map section

* fix shutdown bug

* fix graph update config slightly

* avoid pgmo warning and move print

* match khronos verbosity control for projective integrator

* (wip) start on graph interlayer edge code for arbitrary layers

* start testing graph connector

* clean up and fix connector

* fix active flags

* add graph connector to frontend

* fix active and config issues

* document packet fields and methods

* add default graph update config for objects

* fix active window sink parsing

* clean up printInfo and sink printing

* maybe avoid yamlcpp warning

* fix compilation error

+++ src/reconstruction/mesh_integrator.cpp
@@ -170,7 +170,7 @@ void MeshIntegrator::processExterior(VolumetricMap* map,
   BlockIndex block_index;
   while (index_getter->getNextIndex(block_index)) {
     VLOG(10) << "Extracting exterior for block: " << showIndex(block_index);
-    const auto vps = map->config.voxels_per_side;
+    const auto vps = static_cast<int>(map->config.voxels_per_side);
     VoxelIndex v_idx;
 
     // Max X plane
